% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graficos.R, R/graphs_traceplots.R
\name{graphs_traceplots}
\alias{graphs_traceplots}
\title{An auxiliary function to generate traceplot graphs based on MCMC results for further analysis.}
\usage{
graphs_traceplots(resultados, dados, plot_color, add_trendline, save_file)

graphs_traceplots(resultados, dados, plot_color, add_trendline, save_file)
}
\arguments{
\item{resultados}{A list containing MCMC results, including elements like "S", "burn", "alfa", "mu", "tau.theta", "tau.phi", and "tau.e".}

\item{dados}{A list containing dataset information.}

\item{plot_color}{A character indicating the color for the plot lines.}

\item{add_trendline}{A logical indicating whether to add a trendline to the plots.}

\item{save_file}{A logical indicating whether to save the generated plots to files.}
}
\value{
It generates traceplot graphs based on the provided MCMC results and saves them if specified.

It generates traceplot graphs based on the provided MCMC results and saves them if specified.
}
\description{
An auxiliary function to generate traceplot graphs based on MCMC results for further analysis.

An auxiliary function to generate traceplot graphs based on MCMC results for further analysis.
}
\examples{
# Example usage:
# graphs_traceplots(resultados, dados, plot_color = "blue", add_trendline = TRUE, save_file = TRUE)

# Example usage:
# graphs_traceplots(resultados, dados, plot_color = "blue", add_trendline = TRUE, save_file = TRUE)

}
\references{
NUNES, Letícia. Métodos de Simulação de Dados Geográficos Sintéticos Para Bases Confidenciais. *Dissertação de Mestrado*, [s. l.], 2018.
Disponível em: \url:{http//est.ufmg.br/portal/arquivos/mestrado/dissertacoes/dissertacao_Leticia_Silva_Nunes.pdf}. Acesso em: 2 mar. 2022.

NUNES, Letícia. Métodos de Simulação de Dados Geográficos Sintéticos Para Bases Confidenciais. *Dissertação de Mestrado*, [s. l.], 2018.
Disponível em: \url:{http//est.ufmg.br/portal/arquivos/mestrado/dissertacoes/dissertacao_Leticia_Silva_Nunes.pdf}. Acesso em: 2 mar. 2022.
}
