% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/media_lambda.R
\name{media_lambda}
\alias{media_lambda}
\title{An auxiliary function to create visual representations of lambda-related values based on MCMC results.}
\usage{
media_lambda(results, data, grid, coord, standard_scale, save_file)
}
\arguments{
\item{results}{A list containing MCMC results, including elements like "S", "burn", "lambda", and "media.lambda".}

\item{data}{A list containing dataset information.}

\item{grid}{A numeric value indicating the grid division for plotting.}

\item{coord}{A dataframe containing longitude and latitude coordinates.}

\item{standard_scale}{A logical value indicating whether to standardize color scales in the generated plots.}

\item{save_file}{A logical value indicating whether to save the generated plots as PDF files.}
}
\value{
This function doesn't explicitly return anything. It generates spatial graphs based on lambda-related data and saves them if specified.
}
\description{
It generates spatially-oriented graphs representing lambda values using color scales.
}
\examples{
# Example usage:
# media_lambda(results, data, grid = 10, coord, standard_scale = TRUE, save_file = TRUE)

}
