% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatteplot_origin.R
\name{plot_origin}
\alias{plot_origin}
\title{Scatter Plot for Original Data}
\usage{
plot_origin(coord)
}
\arguments{
\item{coord}{A data frame containing the original coordinates with columns named "lon" and "lat".}
}
\value{
This function creates a scatter plot for the original data.
}
\description{
Scatter Plot for Original Data
}
\examples{
# Example usage:
# coordenadas <- data.frame(
#   lon = c(0.2288, 3.6682, 5.2056, 5.2007, 6.6556, 9.4929),
#   lat = c(0.1211, 2.8985, 4.8331, 4.7690, 6.6016, 9.6699)
# )
# plot_origin(coordenadas)

}
