% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatteplot_syn.R
\name{plot_syn}
\alias{plot_syn}
\title{Scatter Plot for Synthetic Data}
\usage{
plot_syn(coord_syn, n_syn)
}
\arguments{
\item{coord_syn}{A list containing synthetic coordinates.}

\item{n_syn}{An integer specifying the number of synthetic coordinate sets to plot (must be between 4 and 10).}
}
\value{
This function creates scatter plots for the synthetic data.
}
\description{
Scatter Plot for Synthetic Data
}
\examples{
# Example usage:
# dados_sem_coord <- as.data.frame(dados_sem_coord)
# coordenadas <- as.data.frame(coordenadas)
# n_syn <- 5
# coord_restricted_area <- matrix(c(
#   4, 4,
#   4, 5,
#   6, 5,
#   6, 4,
#   4, 4
# ), ncol = 2, byrow = TRUE)
# coordenadas_sinteticas <- syncoordinates(dataset = dados_sem_coord,
#                                          coord = coordenadas,
#                                          grid = 10,
#                                          continuous = 4,
#                                          list_mcmc = mcmc_dados,
#                                          n.syn = n_syn,
#                                          coord_restricted_area = coord_restricted_area)
# plot_syn(coordenadas_sinteticas, n_syn)

}
