% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{A function that generates useful objects for the package from the database}
\usage{
prepare_data(dataset, coord, grid = 10)
}
\arguments{
\item{dataset}{A data frame with all the information except the coordinates}

\item{coord}{An object with two columns indicating the latitude and longitude respectively of the elements in the dataset}

\item{grid}{The grid represents the quantities of divisions that will be made in the location. Bigger the grid, closer the synthetic coordinates are to the real coordinates. With a default result of (grid = 10)}
}
\value{
A list containing useful objects to the mcmc function.
}
\description{
An auxiliary function to generate useful objects for the order functions in this package
Function \code{prepare_data} has the goal of receiving the database of the user to generate important variables that will be used in the MCMC.
and in the end to generate the synthetic coordinates.
In the input, the function receives the parameters: \code{dataset}, \code{coord}, \code{grid}.
}
\references{
NUNES, Letícia. Métodos de Simulação de Dados Geográficos Sintéticos Para Bases Confidenciais. *Dissertação de Mestrado*, [s. l.], 2018.
Disponível em: \url:{http//est.ufmg.br/portal/arquivos/mestrado/dissertacoes/dissertacao_Leticia_Silva_Nunes.pdf}. Acesso em: 2 mar. 2022.
}
